feat(lockboxes): Implement canUnlockAll function
This commit introduces the  function in the  module. The function determines if all boxes in a given list can be unlocked using the keys contained within the boxes. It uses a breadth-first search approach, starting with the keys in the first box and continuing to unlock boxes and gather new keys until there are no more keys or all boxes are unlocked.

Resolves: #123